+++ src-tests/Process/ProcessSetTest.php
@@ -6,6 +6,7 @@ use Graphp\Algorithms\Tree\OutTree;
 use Lmc\Steward\Process\Fixtures\MockOrderStrategy;
 use Lmc\Steward\Publisher\XmlPublisher;
 use Symfony\Component\Console\Output\BufferedOutput;
+use Symfony\Component\Console\Output\Output;
 use Symfony\Component\Process\Process;
 
 class ProcessSetTest extends \PHPUnit_Framework_TestCase
@@ -252,6 +253,55 @@ class ProcessSetTest extends \PHPUnit_Framework_TestCase
         $this->assertNull($preparedProcesses['PreparedTest']->result);
     }
 
+    public function testShouldCountProcessStatuses()
+    {
+        $this->set->add(new Process(''), 'DoneTest1');
+        $this->set->add(new Process(''), 'QueuedTest');
+        $this->set->add(new Process(''), 'PreparedTest');
+        $this->set->add(new Process(''), 'DoneTest2');
+
+        $this->set->setStatus('DoneTest1', ProcessSet::PROCESS_STATUS_DONE);
+        $this->set->setStatus('PreparedTest', ProcessSet::PROCESS_STATUS_PREPARED);
+        $this->set->setStatus('DoneTest2', ProcessSet::PROCESS_STATUS_DONE);
+
+        $this->assertEquals(
+            [
+                ProcessSet::PROCESS_STATUS_PREPARED => 1,
+                ProcessSet::PROCESS_STATUS_QUEUED => 1,
+                ProcessSet::PROCESS_STATUS_DONE => 2,
+            ],
+            $this->set->countStatuses()
+        );
+    }
+
+    public function testShouldCountResultsOfDoneProcesses()
+    {
+        $doneProcessMock = $this->getMockBuilder(Process::class)
+            ->disableOriginalConstructor()
+            ->getMock();
+        $doneProcessMock->expects($this->exactly(4))
+            ->method('getExitCode')
+            ->willReturnOnConsecutiveCalls(0, 255, 2, 0);
+
+        $this->set->add($doneProcessMock, 'DoneTestPassed1');
+        $this->set->add($doneProcessMock, 'DoneTestFatal');
+        $this->set->add($doneProcessMock, 'DoneTestFailed');
+        $this->set->add($doneProcessMock, 'DoneTestPassed2');
+        $this->set->setStatus('DoneTestPassed1', ProcessSet::PROCESS_STATUS_DONE);
+        $this->set->setStatus('DoneTestFatal', ProcessSet::PROCESS_STATUS_DONE);
+        $this->set->setStatus('DoneTestFailed', ProcessSet::PROCESS_STATUS_DONE);
+        $this->set->setStatus('DoneTestPassed2', ProcessSet::PROCESS_STATUS_DONE);
+
+        $this->assertEquals(
+            [
+                ProcessSet::PROCESS_RESULT_PASSED => 2,
+                ProcessSet::PROCESS_RESULT_FAILED => 1,
+                ProcessSet::PROCESS_RESULT_FATAL => 1,
+            ],
+            $this->set->countResults()
+        );
+    }
+
     public function testShouldPublishProcessStatusWhenStatusWasSet()
     {
         $publisherMock = $this->getMockBuilder(XmlPublisher::class)
@@ -292,7 +342,7 @@ class ProcessSetTest extends \PHPUnit_Framework_TestCase
         $processes = $this->set->get(ProcessSet::PROCESS_STATUS_QUEUED);
         $this->assertCount(2, $processes);
 
-        $outputBuffer = new BufferedOutput();
+        $outputBuffer = new BufferedOutput(Output::VERBOSITY_DEBUG);
         // Dequeue process without delay
         $this->set->dequeueProcessesWithoutDelay($outputBuffer);