+++ CHANGELOG.md
@@ -4,7 +4,15 @@
 
 ## Unreleased
 ### Added
-- Provide information about results of finished processes (verbose (`-v`) mode of `run` command must be enabled)
+- Provide information about results of finished processes in output of the `run` command
+
+### Changed
+- Adjust output verbosity for different verbosity levels to improve clarity of the output mainly when running the tests locally.
+    - Use default level if you care only about test results count.
+    - Use verbose mode (`-v`) to see name of failed tests during execution.
+    - Use very verbose mode (`-vv`) to see detailed progress information during execution and also output of failed tests.
+    - Debug level (`-vvv`) should be used when you want to know all available information about the run or when you run the tests on CI server. The tests output is printed incrementally.
+- Set status of timeouted tests as done and their result as failed
 
 ## 1.2.0 - 2016-01-11
 ### Added